finishing up blinker control and adding documentation

+++ vac.h
@@ -23,7 +23,7 @@ public:
          byte _tails,
          byte _brakes,
          byte _heads,
-         byte _blinkerTiming,
+         int _blinkerTiming,
          byte _autoTails,
          byte _debug);
